examples: apply minor updates
including moving common "map_open" function to a helper header.
