+++ examples/CMakeLists.txt
@@ -9,6 +9,8 @@
 #
 add_custom_target(examples)
 
+include_directories(.)
+
 # ----------------------------------------------------------------- #
 ## Setup examples
 # ----------------------------------------------------------------- #
